+++ gluecodium/src/test/resources/smoke/comments/output/dart/lib/src/smoke/internal_class_with_comments.dart
@@ -7,11 +7,10 @@ import 'package:meta/meta.dart';
 /// @nodoc
 @internal
 abstract class InternalClassWithComments {
-
   /// This is definitely internal
   ///
   /// @nodoc
-  void internal_doNothing();
+  void doNothing();
 }
 // InternalClassWithComments "private" section, not exported.
 final _smokeInternalclasswithcommentsRegisterFinalizer = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<
@@ -28,9 +27,8 @@ final _smokeInternalclasswithcommentsReleaseHandle = __lib.catchArgumentError(()
   >('library_smoke_InternalClassWithComments_release_handle'));
 class InternalClassWithComments$Impl extends __lib.NativeBase implements InternalClassWithComments {
   InternalClassWithComments$Impl(Pointer<Void> handle) : super(handle);
-
   @override
-  void internal_doNothing() {
+  void doNothing() {
     final _doNothingFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Void Function(Pointer<Void>, Int32), void Function(Pointer<Void>, int)>('library_smoke_InternalClassWithComments_doNothing'));
     final _handle = this.handle;
     _doNothingFfi(_handle, __lib.LibraryContext.isolateId);