+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Breaking changes:
+  * Functions and properties with `internal` visibility are now unconditionally skipped in generated Dart code.
+
 ## 11.6.0
 Release date: 2022-06-23
 ### Features: